feature/#82/#87/#91/#97 신고하기 기능 및 필터링

+++ src/main/java/com/pd/gilgeorigoreuda/search/controller/SearchController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -21,8 +22,8 @@ public class SearchController {
 
     @GetMapping("/address")
     public ResponseEntity<AddressSearchListResponse> addressAndCategorySearch(
-            @RequestParam(required = false,name = "lat") final Double lat,
-            @RequestParam(required = false,name = "lng") final Double lng,
+            @RequestParam(required = false,name = "lat") final BigDecimal lat,
+            @RequestParam(required = false,name = "lng") final BigDecimal lng,
             @RequestParam(required = false,name = "foodType") final FoodType foodType
     ){
         AddressSearchListResponse response = searchService.getStoreByAddressAndFoodType(lat, lng, foodType);
