+++ src/main/java/com/pd/gilgeorigoreuda/search/repository/SearchRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface SearchRepository extends JpaRepository<Store, Long> {
@@ -14,8 +15,8 @@ public interface SearchRepository extends JpaRepository<Store, Long> {
             "left join fetch s.foodCategories fc " +
             "where POW(s.lat - :lat, 2) + POW(s.lng - :lng, 2) <= :distance " +
             "and (:foodType is null or fc.foodType = :foodType)")
-    List<Store> getStoreByAddressAndFoodType(@Param("lat") final Double lat,
-                                             @Param("lng") final Double lng,
+    List<Store> getStoreByAddressAndFoodType(@Param("lat") final BigDecimal lat,
+                                             @Param("lng") final BigDecimal lng,
                                              @Param("foodType") final FoodType foodType,
                                              @Param("distance") final Double distance);
